Test variables and failing scripts

+++ src/test/java/com/github/gquintana/beepbeep/sql/SqlPipelineBuilderTest.java
@@ -1,5 +1,6 @@
 package com.github.gquintana.beepbeep.sql;
 
+import com.github.gquintana.beepbeep.BeepBeepException;
 import com.github.gquintana.beepbeep.TestConsumer;
 import com.github.gquintana.beepbeep.TestFiles;
 import com.github.gquintana.beepbeep.pipeline.ResultEvent;
@@ -16,15 +17,20 @@ import org.junit.rules.TemporaryFolder;
 import java.io.File;
 import java.io.IOException;
 import java.sql.Connection;
+import java.sql.SQLException;
 import java.sql.Statement;
+import java.util.HashMap;
 import java.util.function.Predicate;
+import java.util.stream.Collectors;
 
 import static com.github.gquintana.beepbeep.sql.TestSqlConnectionProviders.createSqlConnectionProvider;
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.Assert.fail;
 
 public class SqlPipelineBuilderTest {
     @Rule
     public final TemporaryFolder temporaryFolder = new TemporaryFolder();
+    private boolean store = false;
 
     private static void writeResource(ScriptStartEvent e, File scriptFolder) {
         try {
@@ -56,6 +62,7 @@ public class SqlPipelineBuilderTest {
         assertThat(output.events(ScriptStartEvent.class)).hasSize(2);
         assertThat(output.events(ScriptEndEvent.class)).hasSize(2);
         assertThat(output.events(ResultEvent.class)).hasSize(2 + 2 + 2);
+        close(connectionProvider);
     }
 
     @Test
@@ -66,10 +73,10 @@ public class SqlPipelineBuilderTest {
         Predicate<String> resourceFilter = name -> name.startsWith("com/github/gquintana/beepbeep/init/sql/") && name.endsWith(".sql");
         SqlPipelineBuilder pipelineBuilder = new SqlPipelineBuilder()
             .withConnectionProvider(connectionProvider)
-            .withVariable("variable", "value")
             .withScriptStore("beepbeep")
             .withEndConsumer(output)
             .withResourcesScriptScanner(getClass().getClassLoader(), resourceFilter);
+        store = true;
         pipelineBuilder.scan();
         output.clear();
         // When
@@ -80,14 +87,91 @@ public class SqlPipelineBuilderTest {
         close(connectionProvider);
     }
 
-    private static void close(SingleSqlConnectionProvider connectionProvider) throws Exception {
+    @Test
+    public void testConsume_Variable() throws Exception {
+        // Given
+        TestConsumer<ScriptEvent> output = new TestConsumer<>();
+        SingleSqlConnectionProvider connectionProvider = new SingleSqlConnectionProvider(createSqlConnectionProvider());
+        Predicate<String> resourceFilter = name -> name.startsWith("com/github/gquintana/beepbeep/sql/init_vars/") && name.endsWith(".sql");
+        SqlPipelineBuilder pipelineBuilder = new SqlPipelineBuilder()
+            .withConnectionProvider(connectionProvider)
+            .withVariable("person.login", "wile")
+            .withVariable("person.email", "wile.coyote@warnerbros.com")
+            .withEndConsumer(output)
+            .withResourcesScriptScanner(getClass().getClassLoader(), resourceFilter);
+        // When
+        pipelineBuilder.scan();
+        // Then
+        output.assertNoScriptEndFailed();
+        assertThat(output.events).isNotEmpty();
+        assertThat(output.eventStream(ResultEvent.class).map(e -> e.getResult()).collect(Collectors.toSet())).contains("0;wile;wile.coyote@warnerbros.com");
+        close(connectionProvider);
+    }
+
+    @Test
+    public void testConsume_Variables() throws Exception {
+        // Given
+        TestConsumer<ScriptEvent> output = new TestConsumer<>();
+        SingleSqlConnectionProvider connectionProvider = new SingleSqlConnectionProvider(createSqlConnectionProvider());
+        Predicate<String> resourceFilter = name -> name.startsWith("com/github/gquintana/beepbeep/sql/init_vars/") && name.endsWith(".sql");
+        HashMap<String, Object> vars = new HashMap<>();
+        vars.put("person.login", "wile");
+        vars.put("person.email", "wile.coyote@warnerbros.com");
+        SqlPipelineBuilder pipelineBuilder = new SqlPipelineBuilder()
+            .withConnectionProvider(connectionProvider)
+            .withVariables(vars)
+            .withEndConsumer(output)
+            .withResourcesScriptScanner(getClass().getClassLoader(), resourceFilter);
+        // When
+        pipelineBuilder.scan();
+        // Then
+        output.assertNoScriptEndFailed();
+        assertThat(output.events).isNotEmpty();
+        assertThat(output.eventStream(ResultEvent.class).map(e -> e.getResult()).collect(Collectors.toSet())).contains("0;wile;wile.coyote@warnerbros.com");
+        close(connectionProvider);
+    }
+
+    @Test
+    public void testConsume_Fail() throws Exception {
+        // Given
+        TestConsumer<ScriptEvent> output = new TestConsumer<>();
+        SingleSqlConnectionProvider connectionProvider = new SingleSqlConnectionProvider(createSqlConnectionProvider());
+        Predicate<String> resourceFilter = name -> name.startsWith("com/github/gquintana/beepbeep/sql/init_fail/") && name.endsWith(".sql");
+        SqlPipelineBuilder pipelineBuilder = new SqlPipelineBuilder()
+            .withConnectionProvider(connectionProvider)
+            .withEndConsumer(output)
+            .withResourcesScriptScanner(getClass().getClassLoader(), resourceFilter);
+        // When
+        try {
+            pipelineBuilder.scan();
+            fail("Exception expected");
+        } catch (BeepBeepException e) {
+            // Exception expected
+        }
+        // Then
+        assertThat(output.events).hasSize(2);
+        assertThat(output.events(ScriptStartEvent.class)).hasSize(1);
+        assertThat(output.events(ScriptEndEvent.class).get(0).getType()).isEqualTo(ScriptEndEvent.FAIL_TYPE);
+        close(connectionProvider);
+    }
+
+    private void close(SqlConnectionProvider connectionProvider) throws Exception {
         try (Connection connection = connectionProvider.getConnection();
              Statement statement = connection.createStatement()) {
-            statement.execute("DROP TABLE beepbeep");
+            if (store) {
+                statement.execute("DROP TABLE beepbeep");
+            }
+            try {
+                statement.execute("DROP TABLE person");
+            } catch (SQLException e) {
+
+            }
             // When using sequences
             // statement.execute("DROP SEQUENCE beepbeep_seq");
         }
-        connectionProvider.close();
+        if (connectionProvider instanceof SingleSqlConnectionProvider) {
+            ((SingleSqlConnectionProvider) connectionProvider).close();
+        }
     }
 
 }
